Update - 12/04/2025
Novidades:
  - Substituição de alerts padrão do JavaScript por alertas personalizados na tela
  - Implementação de AJAX para:
    - Troca de tema dinâmica
    - Visualização do aplicativo de agendamento
    - Outras funcionalidades
  - Nova tela de visualização de equipamento no aplicativo de agendamento
  - Nova seção de logs recentes na Dashboard
  - Nova tela de visualização de todos os agendamentos na Dashboard
  - Novo campo opcional "número de telefone" para cada usuário
  - Novos tipos de equipamentos adicionados
  - Novo tipo "espaço" substituindo o tipo "sala"

Mudanças:
  - Campos Turma e Observação agora são obrigatórios no aplicativo de agendamento
  - Controladores otimizados:
    - Nova função para chamar outros controladores somente quando necessário
    - Construtores agora recebem apenas $conn como parâmetro
  - Melhorias na UI:
    - Atualização no design do Menu e Footer
    - Páginas remasterizadas: Perfil, Usuários, Equipamentos e Turmas
    - Renovação da interface de Equipamentos

Código:
  - Diversas melhorias, otimizações e pequenas correções no código
  - Migração de todos os estilos <style> e CSS customizado para TailwindCSS para compilação

+++ app/src/handlers/user/delete.php
@@ -2,24 +2,43 @@
 $requiredRoles = ['gestao'];
 require_once "../../bootstrap.php";
 
-$userId = intval($_GET['id']);
-
-if ($userId === $_SESSION['id']) {
-  Navigation::alert("Você não pode deletar seu próprio usuário.", "../../../dashboard/pages/usuarios.php");
-  die();
+if ($_SERVER['REQUEST_METHOD'] !== 'GET') {
+  Navigation::alert('Método inválido', $_SERVER['HTTP_REFERER'], 'error', 'Método Inválido');
+  exit;
 }
 
-$user = $userController->getInfo($userId, 'id', ['name']);
-$classController->handleUserDeletion($userId);
-$success = $userController->delete($userId);
+try {
+  $userId = intval($_GET['id']);
+  if ($userId === $_SESSION['id']) {
+    $_SESSION['alert'][] = [
+      'titulo' => 'Erro',
+      'mensagem' => 'Você não pode deletar seu próprio usuário',
+      'tipo' => 'error'
+    ];
+    Navigation::redirect($_SERVER['HTTP_REFERER']);
+    exit;
+  }
 
-$logger->action(
-  $currentUser['id'],
-  'delete',
-  'users',
-  $id,
-  "Equipamento '" . $user['name'] . "' excluído",
-  Security::getIp()
-);
+  $user = $userController->getInfo($userId, 'id', ['name']);
+  $classController->handleUserDeletion($userId);
+  $success = $userController->delete($userId);
 
-Navigation::redirect($_SERVER['HTTP_REFERER'], true);
+  $logger->action(
+    $currentUser['id'],
+    'delete',
+    'users',
+    $id,
+    "Usuário '" . $user['name'] . "' excluído",
+    Security::getIp()
+  );
+
+  $_SESSION['alert'][] = [
+    'titulo' => 'Sucesso',
+    'mensagem' => 'Usuário ' . $user['name'] . ' deletado com sucesso',
+    'tipo' => 'success'
+  ];
+  Navigation::redirect('../../../dashboard/pages/usuarios.php', true);
+} catch (Exception $e) {
+  Navigation::alert($e->getMessage(), $_SERVER['HTTP_REFERER']);
+  exit;
+}
